+++ Userland/Libraries/CMakeLists.txt
@@ -46,6 +46,7 @@ add_subdirectory(LibPDF)
 add_subdirectory(LibProtocol)
 add_subdirectory(LibPublicSuffix)
 add_subdirectory(LibRegex)
+add_subdirectory(LibRIFF)
 add_subdirectory(LibSanitizer)
 add_subdirectory(LibSoftGPU)
 add_subdirectory(LibSQL)
