+++ Userland/Libraries/LibRIFF/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(SOURCES
+    Types.cpp
+)
+
+serenity_lib(LibRIFF riff)
+target_link_libraries(LibRIFF PRIVATE LibCore)
